requirements.txt
streamlit
beautifulsoup4
requests
python-dotenv
openai
numpy